LayoutTests/platform/gtk/fast/spatial-navigation/snav-unit-overflow-and-scroll-in-direction-expected.txt

Start
End
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS true is true
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
